wip refactoring
